Refactor common API helper methods into one place
`getError` and `checkStatus` have the same implementation for anything
that consumes internal APIs.
