+++ src/common/actions/repositories.js
@@ -1,5 +1,6 @@
 import 'isomorphic-fetch';
 
+import { checkStatus, getError } from '../helpers/api';
 import { conf } from '../helpers/config';
 
 const BASE_URL = conf.get('BASE_URL');
@@ -15,26 +16,6 @@ export function setRepositories(repos) {
     type: SET_REPOSITORIES,
     payload: repos
   };
-}
-
-// TODO bartaz: same as in other actions (share in some helper?)
-export function getError(response, json) {
-  const message = (json.payload && json.payload.message) || response.statusText;
-  const error = new Error(message);
-  error.response = response;
-  error.json = json;
-  return error;
-}
-
-// TODO bartaz: same as in other actions (share in some helper?)
-function checkStatus(response) {
-  if (response.status >= 200 && response.status < 300) {
-    return response;
-  } else {
-    return response.json().then((json) => {
-      throw getError(response, json);
-    });
-  }
 }
 
 export function fetchUserRepositories(pageNumber) {
